:rocket: feat: split codes
