+++ src/Makefile
@@ -1,6 +1,11 @@
-CFLAGS=-g -O0
+CFLAGS=-std=c11 -Wall -g -static
+SRCS=$(wildcard *.c)
+OBJS=$(SRCS:.c=.o)
 
-tim: tim.c
+tim: $(OBJS)
+				$(CC) -o tim $(OBJS) $(LDFLAGS)
+
+$(OBJS): tim.h
 
 test: tim
 				./tim -test
